adding modified survey platform with blueprint format
