--- Data/results/concert_brad_results.json
@@ -1,1 +1,0 @@
-{"IjliMTFmMGI1NDY5MjQyNDRmYjY1NmYzNWIzYmNiMGQ1ZGIxZmEzMmQi.YNypmA.ERhEuhdTVaKLbqN5TCpCxNyJKbk": {"First Name": "test", "Last Name": "test", "Email": "test@test.com", "Verify Email": "test@test.com", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "06/30/21"}, "IjliMTFmMGI1NDY5MjQyNDRmYjY1NmYzNWIzYmNiMGQ1ZGIxZmEzMmQi.YNzJlQ.IOEy8X_X7pe8XafbH5KBq1QGNNc": {"First Name": "test", "Last Name": "test", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Email": "test", "Verify Email": "test", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "06/30/21"}, "IjliMTFmMGI1NDY5MjQyNDRmYjY1NmYzNWIzYmNiMGQ1ZGIxZmEzMmQi.YNzKsQ.t1fgpo3W3nTBZ4FGy7jz5pNbVSY": {"First Name": "test", "Last Name": "test", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Email": "test", "Verify Email": "test", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "06/30/21"}, "IjliMTFmMGI1NDY5MjQyNDRmYjY1NmYzNWIzYmNiMGQ1ZGIxZmEzMmQi.YNzLWw.6ZkgbGmhC0ZaP_afsa3sfi4sC6Q": {"First Name": "test", "Last Name": "test", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Email": "test", "Verify Email": "test", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "06/30/21"}, "IjliMTFmMGI1NDY5MjQyNDRmYjY1NmYzNWIzYmNiMGQ1ZGIxZmEzMmQi.YNzaqg.dM1pqNjIlnOZrTIUbkg_NsP-rNs": {"First Name": "test", "Last Name": "test", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Email": "test", "Verify Email": "test@test.com", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "06/30/21"}, "IjliMTFmMGI1NDY5MjQyNDRmYjY1NmYzNWIzYmNiMGQ1ZGIxZmEzMmQi.YNza0A.S9WmVcbx4POKIYwrJ1HscpAMCzw": {"First Name": "test", "Last Name": "test", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Email": "test", "Verify Email": "test", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "06/30/21"}, "IjliMTFmMGI1NDY5MjQyNDRmYjY1NmYzNWIzYmNiMGQ1ZGIxZmEzMmQi.YNzd4g.YZpxJNwX53JgF5Ca75l8deUZdJU": {"First Name": "test", "Last Name": "test", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Email": "test", "Verify Email": "test", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "06/30/21"}, "IjliMTFmMGI1NDY5MjQyNDRmYjY1NmYzNWIzYmNiMGQ1ZGIxZmEzMmQi.YNzfVQ.hyb_jqFPQ6Z72jkfnMfJBm54WCA": {"First Name": "test", "Last Name": "test", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Email": "test", "Verify Email": "test@test.com", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "06/30/21"}, "IjE2YWViNjM4ZWY1NzllZTIwMWRiNjY3ZWVmMWMwNDZlN2E2M2EwNmQi.YN3Oig.78qIEQEKxvyV3uk6huWBIReHiC8": {"First Name": "test", "Last Name": "test", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Email": "test", "Verify Email": "test", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "07/01/21"}, "IjE2YWViNjM4ZWY1NzllZTIwMWRiNjY3ZWVmMWMwNDZlN2E2M2EwNmQi.YN3QJA.G3g4oqkbbqv0SIKgNmVD4QGKoRg": {"First Name": "test", "Last Name": "test", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Email": "test@test.com", "Verify Email": "test@test.com", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "07/01/21"}, "IjE2YWViNjM4ZWY1NzllZTIwMWRiNjY3ZWVmMWMwNDZlN2E2M2EwNmQi.YN3fiw.x1JloSc8dUR6XSYYOhdORF3LRrs": {"First Name": "test", "Last Name": "test", "Department": "test", "Supervisor": "test", "Email": "test@test.com", "Verify Email": "test@test.com", "Phone": "5", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "07/01/21"}, "IjE2YWViNjM4ZWY1NzllZTIwMWRiNjY3ZWVmMWMwNDZlN2E2M2EwNmQi.YN3frA.48RSsbpmGaE9AlwArOKDd5PZsBU": {"First Name": "test", "Last Name": "test", "Department": "test", "Supervisor": "test", "Email": "test@test.com", "Verify Email": "test@test.com", "Phone": "abc", "Location": "Administration", "Shift": "Admin", "submit": "Submit", "date": "07/01/21"}}
