--- Data/results/custom_metadata_results.json
@@ -1,1 +1,0 @@
-{}
