--- Data/results/data.csv
@@ -1,3 +1,0 @@
-firstName,lastName,email,confirmEmail,phone,location,department,supervisor,shift,submit
-test,test,test@test.com,test@test.com,5555555555,Administraton,test,test,Admin,Submit
-Kayla,Hamilton,kayla.hamilton@daimler.com,kayla.hamilton@daimler.com,5555555555,Administraton,Legal,Steven,Admin,Submit
