--- Data/results/data.json
@@ -1,1 +1,0 @@
-{"IjE1NjViN2RhZWU4OGI5NDE4NmVjMzAzZjE1NTVkM2QwZDM4NzcyNTAi.YMoLgw.8-Nbm4OxXh9cmFLi26GLdAjSRDA": {"firstName": "test", "lastName": "test", "email": "test@test.com", "confirmEmail": "test@test.com", "phone": "5555555555", "location": "Administraton", "department": "test", "supervisor": "test", "shift": "Admin", "submit": "Submit"}, "IjE1NjViN2RhZWU4OGI5NDE4NmVjMzAzZjE1NTVkM2QwZDM4NzcyNTAi.YMoLnw.OyVnZjZAytwSEf--ezRcHq5m3Gw": {"firstName": "Kayla", "lastName": "Hamilton", "email": "kayla.hamilton@daimler.com", "confirmEmail": "kayla.hamilton@daimler.com", "phone": "5555555555", "location": "Administraton", "department": "Legal", "supervisor": "Steven", "shift": "Admin", "submit": "Submit"}}
