--- Data/results/survey_response.csv
@@ -1,3 +1,0 @@
-First Name,Last Name,Email,Verify Email,Phone,Department,Supervisor,Location,Shift,submit
-test,test,test@test.com,test@test.com,5555555555,test,test,Administraton,Admin,Submit
-test1,test2,test1.test2@test.com,test1.test2@test.com,5555555555,test,test3,Bibb County - GSP,A,Submit
