--- Data/results/survey_response.json
@@ -1,1 +1,0 @@
-{"Ijg1ZDg3ZDQwMDdjODNjMDc2NmM0OWNkYzU3NGYzNTIyZWY3NDM5OWYi.YMkFvg.IORoCoXTuL1cM88e0WTRVDjUczk": {"First Name": "test", "Last Name": "test", "Email": "test@test.com", "Verify Email": "test@test.com", "Phone": "5555555555", "Department": "test", "Supervisor": "test", "Location": "Administraton", "Shift": "Admin", "submit": "Submit"}, "IjE1NjViN2RhZWU4OGI5NDE4NmVjMzAzZjE1NTVkM2QwZDM4NzcyNTAi.YMoH3g.C_S1kXNDu9zQNY-hqTSRrGlZBeA": {"First Name": "test1", "Last Name": "test2", "Email": "test1.test2@test.com", "Verify Email": "test1.test2@test.com", "Phone": "5555555555", "Department": "test", "Supervisor": "test3", "Location": "Bibb County - GSP", "Shift": "A", "submit": "Submit"}}
