--- Data/results/test_survey_results.csv
@@ -1,2 +1,0 @@
-Name,Email,T-Shirt Size,Color,submit
-Kayla,test@test.com,M,Blue,Submit
