--- Data/results/test_survey_results.json
@@ -1,1 +1,0 @@
-{}
