+++ flask_wtforms_tutorial/__init__.py
@@ -1,15 +1,40 @@
 """Initialize app."""
 from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from flask_login import LoginManager
 
+db = SQLAlchemy()
 
 def create_app():
     """Construct the core flask_wtforms_tutorial."""
     app = Flask(__name__, instance_relative_config=False)
     app.config.from_object("config.Config")
+    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///db.sqlite'
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+    
+    db.init_app(app)
+    
+    login_manager = LoginManager()
+    login_manager.login_view = 'auth.login'
+    login_manager.init_app(app)
 
+    from .auth.models import User
+    
+    @login_manager.user_loader
+    def load_user(user_id):
+        # since the user_id is just the primary key of our user table, use it in the query for the user
+        return User.query.get(int(user_id))
 
     with app.app_context():
         # Import parts of our flask_wtforms_tutorial
-        from . import routes
+        
+        from .auth import auth
+        app.register_blueprint(auth.auth)
+        
+        from .admin import admin
+        app.register_blueprint(admin.admin_bp, url_prefix='/admin')
+        
+        from .surveys import surveys
+        app.register_blueprint(surveys.survey_bp)
         
         return app
