+++ flask_wtforms_tutorial/admin/templates/admin/update.jinja2
@@ -1,0 +1,27 @@
+{% extends 'layout.jinja2' %}
+
+{% block styles %}
+
+    <link
+        rel="stylesheet"
+        href="{{ url_for('static', filename='css/forms.css') }}"
+        type="text/css"
+    />
+
+{% endblock %}
+
+{% block content %}
+
+    <div class="form-wrapper">
+
+      <h1> {{ title }}	
+      <h2>Would you like to update the {{ file }} survey?</h2>
+
+      <form method=post enctype=multipart/form-data>
+        <input type=submit name=Cancel value=Cancel>
+	<input type=submit name=Continue value=Continue>
+      </form>
+
+    </div>
+
+{% endblock %}
