+++ flask_wtforms_tutorial/admin/templates/delete.jinja2
