+++ flask_wtforms_tutorial/admin/templates/upload.jinja2
