+++ flask_wtforms_tutorial/auth/models.py
@@ -1,0 +1,8 @@
+from flask_login import UserMixin
+from .. import db
+
+class User(db.Model, UserMixin):
+    id = db.Column(db.Integer, primary_key=True) # primary keys are required by SQLAlchemy
+    email = db.Column(db.String(100), unique=True)
+    password = db.Column(db.String(100))
+    name = db.Column(db.String(1000))
