--- flask_wtforms_tutorial/data/data.json
@@ -1,1 +1,0 @@
-[]
