+++ flask_wtforms_tutorial/surveys/templates/surveys/survey.jinja2
@@ -15,7 +15,7 @@
     <div class="form-wrapper">
       <h2 class="title">{{ title }}</h2>
 
-      <form method="POST" action="{{ url_for('create_survey', name=name) }}">
+      <form method="POST" action="{{ url_for('survey_bp.create_survey', name=name) }}">
           {{ form.csrf_token }}
 
           {% for text_form in form.text_entries %}
@@ -65,6 +65,20 @@
                 {% if select_form.select.errors %}
                   <ul class="errors">
                     {% for error in select_form.select.errors %}
+                      <li class="error">
+                        <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
+                      </li>
+                    {% endfor %}
+                  </ul>
+                {% endif %}
+              </fieldset>
+          {% endfor %}
+
+          {% for multi_form in form.multi_entries %}
+              <fieldset class="form-field"> {{multi_form.multi.label}} {{multi_form.multi}}
+                {% if multi_form.multi.errors %}
+                  <ul class="errors">
+                    {% for error in multi_form.multi.errors %}
                       <li class="error">
                         <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
                       </li>
