--- flask_wtforms_tutorial/templates/contact.jinja2
@@ -1,58 +1,0 @@
-{% extends 'layout.jinja2' %}
-
-{% block styles %}
-    <link
-            rel="stylesheet"
-            href="{{ url_for('static', filename='css/forms.css') }}"
-            type="text/css"
-    />
-{% endblock %}
-
-{% block content %}
-
-    <div class="form-wrapper">
-      <h2 class="title">Contact</h2>
-      <form method="POST" action="{{ url_for('contact') }}">
-        {{ form.csrf_token }}
-
-        <fieldset class="form-field">{{ form.name.label }} {{ form.name(size=20) }}
-          {% if form.name.errors %}
-            <ul class="errors">
-                {% for error in form.email.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-          {% endif %}
-        </fieldset>
-
-        <fieldset class="form-field">{{ form.email.label }} {{ form.email }}
-          {% if form.email.errors %}
-            <ul class="errors">
-                {% for error in form.email.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-          {% endif %}
-        </fieldset>
-
-        <fieldset class="form-field">{{ form.body.label }} {{ form.body }}
-          {% if form.body.errors %}
-            <ul class="errors">
-                {% for error in form.email.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-          {% endif %}
-        </fieldset>
-
-        {{ form.submit }}
-      </form>
-    </div>
-
-{% endblock %}
