--- flask_wtforms_tutorial/templates/signup.jinja2
@@ -1,136 +1,0 @@
-{% extends 'layout.jinja2' %}
-
-{% block styles %}
-
-    <link
-        rel="stylesheet"
-        href="{{ url_for('static', filename='css/forms.css') }}"
-        type="text/css"
-    />
-
-{% endblock %}
-
-{% block content %}
-
-    <div class="form-wrapper">
-      <h2 class="title">Giveaway Signup</h2>
-
-      <form method="POST" action="{{ url_for('signup') }}">
-          {{ form.csrf_token }}
-
-          <fieldset class="form-field">{{ form.firstName.label }} {{ form.firstName }}
-            {% if form.firstName.errors %}
-              <ul class="errors">
-                {% for error in form.firstName.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-            {% endif %}
-          </fieldset>
-
-          <fieldset class="form-field">{{ form.lastName.label }} {{ form.lastName }}
-            {% if form.lastName.errors %}
-              <ul class="errors">
-                {% for error in form.lastName.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-            {% endif %}
-          </fieldset>
-
-          <fieldset class="form-field">{{ form.email.label }} {{ form.email }}
-            {% if form.email.errors %}
-              <ul class="errors">
-                {% for error in form.email.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-            {% endif %}
-          </fieldset>
-
-          <fieldset class="form-field">{{ form.confirmEmail.label }} {{ form.confirmEmail }}
-            {% if form.confirmEmail.errors %}
-              <ul class="errors">
-                {% for error in form.confirmEmail.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-            {% endif %}
-          </fieldset>
-
-          <fieldset class="form-field">{{ form.phone.label }} {{ form.phone }}
-            {% if form.phone.errors %}
-              <ul class="errors">
-                {% for error in form.phone.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-            {% endif %}
-          </fieldset>
-
-          <fieldset class="form-field">{{ form.location.label }} {{ form.location }}
-            <i class="fas fa-caret-down"></i>
-            {% if form.location.errors %}
-              <ul class="errors">
-                {% for error in form.location.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-            {% endif %}
-          </fieldset>
-
-          <fieldset class="form-field">{{ form.department.label }} {{ form.department }}
-            {% if form.department.errors %}
-              <ul class="errors">
-                {% for error in form.department.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-            {% endif %}
-          </fieldset>
-
-          <fieldset class="form-field">{{ form.supervisor.label }} {{ form.supervisor }}
-            {% if form.supervisor.errors %}
-              <ul class="errors">
-                {% for error in form.supervisor.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-            {% endif %}
-          </fieldset>
-
-          <fieldset class="form-field">{{ form.shift.label }} {{ form.shift }}
-            <i class="fas fa-caret-down"></i>
-            {% if form.shift.errors %}
-              <ul class="errors">
-                {% for error in form.shift.errors %}
-                  <li class="error">
-                      <i class="fas fa-exclamation"></i> <span class="error-message">{{ error }}</span>
-                  </li>
-                {% endfor %}
-              </ul>
-            {% endif %}
-          </fieldset>
-
-          {{ form.submit }}
-
-      </form>
-    </div>
-
-{% endblock %}
