+++ wsgi.py
@@ -1,8 +1,8 @@
 """Application entry point."""
-from flask_wtforms_tutorial import create_app
+from mbusi_surveys import create_app
 
 app = create_app()
 
 
 if __name__ == "__main__":
-    app.run(host="0.0.0.0", debug=True, threading=False)
+    app.run(host="0.0.0.0", debug=True)
